fix(Store, Image): lógica para apenas os usuários proprietários e admins poderem atualizar as informações de suas lojas

+++ server/src/controllers/Image.controller.ts
@@ -1,10 +1,11 @@
-import { Body, Controller, Get, Param, Post } from '@nestjs/common';
+import { Body, Controller, Get, Param, Post, Req } from '@nestjs/common';
 import { IsAdminGuard } from 'infra/guards/Roles/is-admin.guard';
 import { UseGuards } from '@nestjs/common';
 import { JwtAccessAuthGuard } from 'infra/guards/Auth/jwt-access.guard';
 import { ImageService } from 'services/Image.service';
 import { Image } from 'domain/entities/Image.entity';
 import { CreateImageDto, UpdateImageDto } from 'domain/dto/Image.dto';
+import { Request } from 'express';
 
 @Controller('api/images')
 export class ImageController {
@@ -29,15 +30,42 @@ export class ImageController {
   @UseGuards(JwtAccessAuthGuard, IsAdminGuard)
   @Post('/update/:id')
   async update(
+    @Req() req: Request,
     @Param('id') id: string,
     @Body() imageDto: UpdateImageDto,
   ): Promise<Image> {
-    return await this.imageService.update(+id, imageDto);
+    let ownerIsCurrentUser: boolean = true;
+    if (!req.user['isAdmin']) {
+      try {
+        ownerIsCurrentUser = await this.imageService.assureOwnerIsCurrentUser(
+          +id,
+          req.user['sub'],
+        );
+      } catch (err) {
+        throw err;
+      }
+    }
+    if (ownerIsCurrentUser) {
+      return await this.imageService.update(+id, imageDto);
+    }
   }
 
   @UseGuards(JwtAccessAuthGuard, IsAdminGuard)
   @Post('/delete/:id')
-  async delete(@Param('id') id: string): Promise<Image> {
-    return await this.imageService.delete(+id);
+  async delete(@Req() req: Request, @Param('id') id: string): Promise<Image> {
+    let ownerIsCurrentUser: boolean = true;
+    if (!req.user['isAdmin']) {
+      try {
+        ownerIsCurrentUser = await this.imageService.assureOwnerIsCurrentUser(
+          +id,
+          req.user['sub'],
+        );
+      } catch (err) {
+        throw err;
+      }
+    }
+    if (ownerIsCurrentUser) {
+      return await this.imageService.delete(+id);
+    }
   }
 }
